Refactor Cleanup
